refactor: product query 분리
